Some reorganization, everything relating to code is now under Code/* and old releases are under an old releases folder, any prototyping solutions are under Prototypes, all test file are under Testbed, testing solutions are under the TestingSoultions folder.
Under Code you will find DatabaseViewer, which is latest DB viewer project, EngineScannerDLL which is the latest C++ engine scanner project, and SSNScanner which is the latest GUI interface and scanner project. There are also folders for 3.5_DLLs and Components which are all 3rd party tools. I am not sure if the previous team has modified those source files from the original vendor